+++ x.rb
@@ -7,3 +7,4 @@ require_relative 'x/xapi'
 require_relative 'x/track'
 require_relative 'x/docs'
 require_relative 'x/problem'
+require_relative 'x/readme'
